requirements.txt
aniso8601==10.0.1
appdirs==1.4.4
attrs==25.4.0
Authlib==1.6.4
Automat==25.4.16
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
constantly==23.10.4
-e git+https://github.com/DyanielCX/FireZWall-Adaptive-Firewall-FYP-Project.git@c2d87acddb91532fb42f1f10e636bc9045b6f44f#egg=cowrie&subdirectory=cowrie
cryptography==46.0.2
Flask==3.1.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multipart==1.3.0
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
tftpy==0.8.6
treq==25.5.0
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.interface==8.0.1
